Working branch (#1)
Fixes an issue where workflow names would be truncated and cause the action to fail.
Root cause: When the line length of the job name + matrix inputs exceeds 97 characters in GHA. The name is truncated and an ellipsis is appended. This caused the action to fail.
Fix: 
Truncate the name just like GHA does. 

Signed-off-by: Collin Charvat <c3charvat@gmail.com>